Draft: Release 1.6

See merge request vng/diwi!629

+++ backend/src/main/java/nl/vng/diwi/resources/ProjectsResource.java
@@ -64,12 +64,15 @@ import nl.vng.diwi.models.ProjectUpdateModel;
 import nl.vng.diwi.models.ProjectUpdateModel.ProjectProperty;
 import nl.vng.diwi.models.SelectModel;
 import nl.vng.diwi.models.SingleValueOrRangeModel;
+import nl.vng.diwi.models.UserGroupUserModel;
 import nl.vng.diwi.models.superclasses.ProjectCreateSnapshotModel;
 import nl.vng.diwi.models.superclasses.ProjectMinimalSnapshotModel;
 import nl.vng.diwi.rest.VngBadRequestException;
+import nl.vng.diwi.rest.VngNotAllowedException;
 import nl.vng.diwi.rest.VngNotFoundException;
 import nl.vng.diwi.rest.VngServerErrorException;
 import nl.vng.diwi.security.LoggedUser;
+import nl.vng.diwi.security.UserAction;
 import nl.vng.diwi.security.UserActionConstants;
 import nl.vng.diwi.services.ExcelImportService;
 import nl.vng.diwi.services.GeoJsonImportService;
@@ -126,6 +129,23 @@ public class ProjectsResource {
             throw new VngBadRequestException(validationError);
         }
 
+        boolean userIsOwner = false;
+        for (UserGroupModel group : projectSnapshotModel.getProjectOwners()) {
+            List<UUID> userIds = repo.getUsergroupDAO().getUserGroupUsers(group.getUuid()).stream().map(UserGroupUserModel::getUuid).toList();
+            if (userIds.contains(loggedUser.getUuid())) {
+                userIsOwner = true;
+                break;
+            }
+        }
+        if (!userIsOwner) {
+            if (!loggedUser.getRole().allowedActions.contains(UserAction.EDIT_ALL_PROJECTS)) {
+                throw new VngBadRequestException("Cannot create projects that you are not the owner of.");
+            }
+            if (projectSnapshotModel.getConfidentialityLevel() == Confidentiality.PRIVATE) {
+                throw new VngBadRequestException("Cannot create private projects that you are not the owner of.");
+            }
+        }
+
         ZonedDateTime now = ZonedDateTime.now();
 
         try (AutoCloseTransaction transaction = repo.beginTransaction()) {
@@ -151,9 +171,12 @@ public class ProjectsResource {
 
     @DELETE
     @Path("/{id}")
-    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS})
-    public void deleteProject(ContainerRequestContext requestContext, @PathParam("id") UUID projectUuid) throws VngNotFoundException {
+    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS, UserActionConstants.EDIT_ALL_PROJECTS})
+    public void deleteProject(ContainerRequestContext requestContext, @PathParam("id") UUID projectUuid) throws VngNotFoundException, VngNotAllowedException {
         var loggedUser = (LoggedUser) requestContext.getProperty("loggedUser");
+
+        projectService.checkProjectEditPermission(repo, projectUuid, loggedUser);
+
         try (AutoCloseTransaction transaction = repo.beginTransaction()) {
             projectService.deleteProject(repo, projectUuid, loggedUser.getUuid());
             transaction.commit();
@@ -233,11 +256,11 @@ public class ProjectsResource {
 
     @PUT
     @Path("/{id}/customproperties")
-    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS})
+    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS, UserActionConstants.EDIT_ALL_PROJECTS})
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public List<ProjectHouseblockCustomPropertyModel> updateProjectCustomProperty(@Context LoggedUser loggedUser, @PathParam("id") UUID projectUuid, ProjectHouseblockCustomPropertyModel projectCPUpdateModel)
-        throws VngNotFoundException, VngBadRequestException, VngServerErrorException {
+        throws VngNotFoundException, VngBadRequestException, VngServerErrorException, VngNotAllowedException {
 
         if (projectCPUpdateModel.getCustomPropertyId() == null){
             throw new VngBadRequestException("Custom property id must be set.");
@@ -252,6 +275,8 @@ public class ProjectsResource {
         if (dbCP.getType() != PropertyKind.CUSTOM) {
             throw new VngBadRequestException("Not a custom property.");
         }
+
+        projectService.checkProjectEditPermission(repo, projectUuid, loggedUser);
 
         LocalDate updateDate = LocalDate.now();
 
@@ -354,15 +379,17 @@ public class ProjectsResource {
 
     @POST
     @Path("/{id}/plots")
-    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS})
+    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS, UserActionConstants.EDIT_ALL_PROJECTS})
     @Consumes(MediaType.APPLICATION_JSON)
-    public void setProjectPlots(@Context LoggedUser loggedUser, @PathParam("id") UUID projectUuid, List<PlotModel> plots) throws VngNotFoundException, VngBadRequestException {
+    public void setProjectPlots(@Context LoggedUser loggedUser, @PathParam("id") UUID projectUuid, List<PlotModel> plots) throws VngNotFoundException, VngBadRequestException, VngNotAllowedException {
 
         Project project = projectService.getCurrentProject(repo, projectUuid);
 
         if (project == null) {
             throw new VngNotFoundException("project not found");
         }
+
+        projectService.checkProjectEditPermission(repo, projectUuid, loggedUser);
 
         for (var plot : plots) {
             String validationError = plot.validate();
@@ -388,11 +415,11 @@ public class ProjectsResource {
 
     @POST
     @Path("/update")
-    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS})
+    @RolesAllowed({UserActionConstants.EDIT_OWN_PROJECTS, UserActionConstants.EDIT_ALL_PROJECTS})
     @Produces(MediaType.APPLICATION_JSON)
     @Consumes(MediaType.APPLICATION_JSON)
     public ProjectSnapshotModel updateProjectSnapshot(@Context LoggedUser loggedUser, ProjectSnapshotModel projectSnapshotModelToUpdate)
-            throws VngNotFoundException, VngBadRequestException, VngServerErrorException {
+        throws VngNotFoundException, VngBadRequestException, VngServerErrorException, VngNotAllowedException {
 
         UUID projectUuid = projectSnapshotModelToUpdate.getProjectId();
         Project project = projectService.getCurrentProject(repo, projectUuid);
@@ -401,6 +428,17 @@ public class ProjectsResource {
         }
 
         ProjectSnapshotModel projectSnapshotModelCurrent = projectService.getProjectSnapshot(repo, projectUuid, loggedUser);
+        projectService.checkProjectEditPermission(repo, projectUuid, projectSnapshotModelCurrent, loggedUser);
+
+        if (projectSnapshotModelToUpdate.getConfidentialityLevel() == Confidentiality.PRIVATE) {
+            List<UUID> newGroupIds = projectSnapshotModelToUpdate.getProjectOwners().stream().map(UserGroupModel::getUuid).toList();
+            List<UUID> newOwnerIds = repo.getUsergroupDAO().getUserGroupsUsers(newGroupIds).stream().map(UserGroupUserModel::getUuid).toList();
+            List<UUID> currentOwnerIds = new ArrayList<>();
+            projectSnapshotModelCurrent.getProjectOwners().forEach(ug -> currentOwnerIds.addAll(ug.getUsers().stream().map(UserGroupUserModel::getUuid).toList()));
+            if (!newOwnerIds.contains(loggedUser.getUuid()) && currentOwnerIds.contains(loggedUser.getUuid())) {
+                throw new VngBadRequestException("Cannot remove yourself as owner of a private project.");
+            }
+        }
 
         List<ProjectUpdateModel> projectUpdateModelList = new ArrayList<>();
         for (ProjectUpdateModel.ProjectProperty projectProperty : ProjectUpdateModel.ProjectProperty.values()) {
@@ -471,16 +509,24 @@ public class ProjectsResource {
                 if (toUpdateOwnersUuids.isEmpty()) {
                     throw new VngBadRequestException("Missing project owners property");
                 }
-                currentOwnersUuids.forEach(uuid -> {
+                for (UUID uuid : currentOwnersUuids) {
                     if (!toUpdateOwnersUuids.contains(uuid)) {
-                        projectUpdateModelList.add(new ProjectUpdateModel(ProjectProperty.projectOwners, null, uuid));
+                        if (!loggedUser.getRole().allowedActions.contains(UserAction.VIEW_GROUPS)) {
+                            throw new VngBadRequestException("User does not have permission to change project user-groups.");
+                        } else {
+                            projectUpdateModelList.add(new ProjectUpdateModel(ProjectProperty.projectOwners, null, uuid));
+                        }
                     }
-                });
-                toUpdateOwnersUuids.forEach(uuid -> {
+                }
+                for (UUID uuid : toUpdateOwnersUuids) {
                     if (!currentOwnersUuids.contains(uuid)) {
-                        projectUpdateModelList.add(new ProjectUpdateModel(ProjectProperty.projectOwners, uuid, null));
+                        if (!loggedUser.getRole().allowedActions.contains(UserAction.VIEW_GROUPS)) {
+                            throw new VngBadRequestException("User does not have permission to change project user-groups.");
+                        } else {
+                            projectUpdateModelList.add(new ProjectUpdateModel(ProjectProperty.projectOwners, uuid, null));
+                        }
                     }
-                });
+                }
             }
             case projectPhase -> {
                 if (!Objects.equals(projectSnapshotModelToUpdate.getProjectPhase(), projectSnapshotModelCurrent.getProjectPhase())) {
@@ -548,7 +594,12 @@ public class ProjectsResource {
             }
         }
 
-        return projectService.getProjectSnapshot(repo, projectUuid, loggedUser);
+        try {
+            return projectService.getProjectSnapshot(repo, projectUuid, loggedUser);
+        } catch (VngNotFoundException ex) {
+            //it can happen that the user does not have access to the project anymore after the update
+            return null;
+        }
     }
 
     private void updateProjectProperty(Project project, ProjectUpdateModel projectUpdateModel, LoggedUser loggedUser, LocalDate updateDate, ZonedDateTime changeDate)
